[Pipelining] Refactor common utils from test_pp_dp
Split test_pp_dp into pp_ddp and pp_fsdp so its a bit more
concise and easier to add CP to the FSDP one.

Realize that 'use_new_runtime' parametrization was not even being used,
removing it saves a bunch of test time. We should migrate schedules to
the new runtime and have them be covered that way.  (And
test_schedule*.py are testing new runtime too).

ghstack-source-id: e2afa673fca195292116a39dfa6d7051ede45beb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/144596