inductor_config_logging: Don't drop keys
This bit me while I was trying to debug some trace issues.
In general this config is already quite large when dumping, so adding
more fields doesn't make it significantly worse.

Also a number of the items we are type checking for (except the test
configs), don't even show up. Primarily this will help us when debugging
rocm, halide, and trace configs.

ghstack-source-id: ff5779a331e9499a55de7a1dacd769fb0d953f3a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/144700